chore: extract user and license seeding
